+++ cached_network_image/example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,12 @@
 import FlutterMacOS
 import Foundation
 
+import path_provider_foundation
+import sqflite
+import url_launcher_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
+  SqflitePlugin.register(with: registry.registrar(forPlugin: "SqflitePlugin"))
+  UrlLauncherPlugin.register(with: registry.registrar(forPlugin: "UrlLauncherPlugin"))
 }