+++ cached_network_image_platform_interface/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [2.0.1] - TBD
 * Specify types
-* Add ErrorListener
+* Add error to ErrorListener
 * Update example
+* Use `loadImage` instead `loadBuffer` and `load`
 
 ## [2.0.0] - 2022-08-31
 * Added loadBufferAsync for Flutter 3.3